Alert refactor
