+++ ImageFeed/Alert/AlertModel.swift
@@ -1,0 +1,18 @@
+//
+//  Alert.swift
+//  ImageFeed
+//
+//  Created by Vitaly Lobov on 06.12.2024.
+//
+
+import UIKit
+
+struct AlertModel {
+    let title: String
+    let message: String
+    let buttonText: String
+    let completion: (() -> Void)?
+    var secondButtonText: String? = nil
+    var secondCompletion: (() -> Void)? = {}
+}
+
