+++ ImageFeed/Alert/AlertPresenter.swift
@@ -1,20 +1,10 @@
 //
-//  Alert.swift
+//  AlertPresenter.swift
 //  ImageFeed
 //
-//  Created by Vitaly Lobov on 06.12.2024.
+//  Created by Vitaly Lobov on 16.12.2024.
 //
-
 import UIKit
-
-struct AlertModel {
-    let title: String
-    let message: String
-    let buttonText: String
-    let completion: (() -> Void)?
-    var secondButtonText: String? = nil
-    var secondCompletion: (() -> Void)? = {}
-}
 
 protocol AlertPresenting: AnyObject {
     func showAlert(for result: AlertModel)
